# Make llm benchmarks query faster