Make faster by adding a projection to the table 
```
PROJECTION benchmark_name_projection
    (
        SELECT *
        ORDER BY 
            repo,
            tupleElement(benchmark, 'name'),
            tupleElement(model, 'name'),
            tupleElement(metric, 'name'),
            timestamp,
            head_branch,
            head_sha,
            workflow_id,
            job_id,
            servicelab_experiment_id,
            servicelab_trial_id
    ),
```
For some reason, ClickHouse won't use the projection unless t uses `has` instead of `in` and all the other tuple elements are in quotation marks 

We should consider changing the order by key to be this instead since I can't imagine anyone not filtering by benchmark name

Perf results:
```
+------+----------+-----------+-------------+---------------+----------+-----------+------------+--------------+
| Test | Avg Time | Base Time | Time Change | % Time Change | Avg Mem  |  Base Mem | Mem Change | % Mem Change |
+------+----------+-----------+-------------+---------------+----------+-----------+------------+--------------+
|  0   |   131    |   33300   |    -33169   |      -100     | 57215999 | 335780375 | -278564376 |     -83      |
+------+----------+-----------+-------------+---------------+----------+-----------+------------+--------------+
```

Also added a test

Checked results were the same after sorting

Also a change to the script to compare queries that makes it easier to compare results that are different by piping them to files that can be diffed later