@huydhn 2025-02-05
I think we should also add some more key to the projection if possible.  They are the keys that I should have added when creating the table.

```
PROJECTION benchmark_name_projection
    (
        SELECT *
        ORDER BY 
            repo,
            tupleElement(benchmark, 'name'),
            tupleElement(model, 'name'),
            tupleElement(metric, 'name'),
            timestamp,
            head_branch,
            head_sha,
            workflow_id,
            job_id,
            servicelab_experiment_id,
            servicelab_trial_id
    ),
```