@huydhn 2025-02-05
Also, could you try update the two queries?

* https://github.com/pytorch/test-infra/tree/main/torchci/clickhouse_queries/oss_ci_benchmark_branches
* and https://github.com/pytorch/test-infra/tree/main/torchci/clickhouse_queries/oss_ci_benchmark_names

I'm currently using a materialize view for them, which I think is slower than this projection approach (when I try to load the page)